refactor: move Logo-Selection to own file
