+++ frontend/src/pages/AddBackupProfilePage.vue
@@ -2,23 +2,17 @@
 import * as backupClient from "../../wailsjs/go/app/BackupClient";
 import * as repoClient from "../../wailsjs/go/app/RepositoryClient";
 import { backupprofile, ent } from "../../wailsjs/go/models";
-import { computed, onMounted, ref, useTemplateRef } from "vue";
+import { computed, ref, useTemplateRef } from "vue";
 import { useRouter } from "vue-router";
 import { rBackupProfilePage, rDashboardPage, withId } from "../router";
 import { showAndLogError } from "../common/error";
-import { useToast } from "vue-toastification";
 import DataSelection from "../components/DataSelection.vue";
 import {
   ArrowRightCircleIcon,
-  BookOpenIcon,
-  BriefcaseIcon,
-  CameraIcon,
   CircleStackIcon,
   ComputerDesktopIcon,
-  EnvelopeIcon,
   FireIcon,
   GlobeEuropeAfricaIcon,
-  HomeIcon,
   PlusCircleIcon
 } from "@heroicons/vue/24/solid";
 import ScheduleSelection from "../components/ScheduleSelection.vue";
@@ -29,6 +23,7 @@ import { formInputClass } from "../common/form";
 import FormField from "../components/common/FormField.vue";
 import { useForm } from "vee-validate";
 import * as yup from "yup";
+import SelectIconModal from "../components/SelectIconModal.vue";
 
 /************
  * Types
@@ -39,12 +34,6 @@ enum Step {
   Schedule = 1,
   Repository = 2,
   Summary = 3,
-}
-
-interface Icon {
-  type: backupprofile.Icon;
-  color: string;
-  html: any;
 }
 
 enum SelectedRepoAction {
@@ -64,35 +53,6 @@ enum SelectedRepoType {
  * Variables
  ************/
 
-const icons: Icon[] = [
-  {
-    type: backupprofile.Icon.home,
-    color: "bg-blue-500 hover:bg-blue-500/50 text-dark dark:text-white",
-    html: HomeIcon
-  },
-  {
-    type: backupprofile.Icon.briefcase,
-    color: "bg-indigo-500 hover:bg-indigo-500/50 text-dark dark:text-white",
-    html: BriefcaseIcon
-  },
-  {
-    type: backupprofile.Icon.book,
-    color: "bg-purple-500 hover:bg-purple-500/50 text-dark dark:text-white",
-    html: BookOpenIcon
-  },
-  {
-    type: backupprofile.Icon.envelope,
-    color: "bg-green-500 hover:bg-green-500/50 text-dark dark:text-white",
-    html: EnvelopeIcon
-  },
-  {
-    type: backupprofile.Icon.camera,
-    color: "bg-yellow-500 hover:bg-yellow-500/50 text-dark dark:text-white",
-    html: CameraIcon
-  },
-  { type: backupprofile.Icon.fire, color: "bg-red-500 hover:bg-red-500/50 text-dark dark:text-white", html: FireIcon }
-];
-
 const router = useRouter();
 const backupProfile = ref<ent.BackupProfile>(ent.BackupProfile.createFrom());
 const currentStep = ref<Step>(Step.SelectData);
@@ -102,9 +62,6 @@ const existingRepos = ref<ent.Repository[]>([]);
 const directorySuggestions = ref<string[]>([]);
 const isBackupPathsValid = ref(false);
 const isExcludePathsValid = ref(true);
-const selectedIcon = ref<Icon>(icons[0]);
-const selectIconModalKey = "select_icon_modal";
-const selectIconModal = useTemplateRef<InstanceType<typeof HTMLDialogElement>>(selectIconModalKey);
 
 const step1Form = useForm({
   validationSchema: yup.object({
@@ -173,17 +130,14 @@ function saveExcludePaths(paths: string[]) {
   backupProfile.value.excludePaths = paths;
 }
 
-function selectIcon(icon: Icon) {
-  backupProfile.value.icon = icon.type;
-  selectedIcon.value = icon;
+function selectIcon(icon: backupprofile.Icon) {
+  backupProfile.value.icon = icon;
 }
 
 async function newBackupProfile() {
   try {
     backupProfile.value = await backupClient.NewBackupProfile();
     directorySuggestions.value = await backupClient.GetDirectorySuggestions();
-
-    selectedIcon.value = icons.find((icon) => icon.type === backupProfile.value.icon) ?? icons[0];
   } catch (error: any) {
     await showAndLogError("Failed to create backup profile", error);
   }
@@ -324,35 +278,7 @@ getExistingRepositories();
         </label>
 
         <!-- Logo -->
-        <button
-          class='btn btn-square'
-          :class='selectedIcon.color'
-          @click='selectIconModal?.showModal()'>
-          <component :is='selectedIcon.html' class='size-8' />
-        </button>
-        <dialog class='modal' autofocus :ref='selectIconModalKey'>
-          <div class='modal-box text-center min-w-fit p-10'>
-            <h3 class='text-lg font-bold pb-6'>Select an icon for this backup profile</h3>
-
-            <form method='dialog'>
-              <div class='grid grid-cols-3 gap-x-12 gap-y-6'>
-                <!-- if there is a button in a form, it will close the modal -->
-                <template v-for='(icon, index) in icons' :key='index'>
-                  <button
-                    class='btn btn-square w-32 h-32'
-                    :class='icon.color'
-                    @click='selectIcon(icon)'
-                  >
-                    <component :is='icon.html' class='size-20' />
-                  </button>
-                </template>
-              </div>
-            </form>
-          </div>
-          <form method='dialog' class='modal-backdrop'>
-            <button>close</button>
-          </form>
-        </dialog>
+        <SelectIconModal :icon=backupProfile.icon @select='selectIcon' />
       </div>
 
       <div class='flex justify-center gap-6 py-10'>
